// src/schemas/ConfigContent.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ConfigContent {
  /**
   * The domain of your HubSpot app
   */
  domain?: string;
  /**
   * The ID of your HubSpot app
   */
  appId?: number;
  /**
   * The cards to provision
   */
  cards?: {
    [k: string]: {
      /**
       * The title of the card
       */
      title: string;
      fetch: {
        /**
         * The target URL of the card
         */
        targetUrl: string;
        /**
         * The object types to fetch
         */
        objectTypes: {
          /**
           * The name of the object type
           */
          name: "contacts" | "companies" | "deals" | "marketing_events" | "tickets";
          /**
           * The properties to fetch
           */
          propertiesToSend: string[];
        }[];
      };
      display?: {
        /**
         * The properties to display
         */
        properties: {
          /**
           * The name of the property
           */
          name: string;
          /**
           * The data type of the property
           */
          dataType: string;
          /**
           * The label of the property
           */
          label: string;
          /**
           * The options of the property
           */
          options?: {
            /**
             * The name of the option
             */
            name: string;
            /**
             * The label of the option
             */
            label: string;
            /**
             * The type of the option
             */
            type: string;
          }[];
        }[];
      };
      actions?: {
        /**
         * The base URLs of the actions
         */
        baseUrls: string[];
      };
    };
  };
  /**
   * The timeline events to provision
   */
  timeline_events?: {
    [k: string]: {
      /**
       * The name of the timeline event
       */
      name: string;
      /**
       * The object type of the timeline event
       */
      objectType: "contacts" | "companies" | "deals" | "tickets";
      /**
       * The header template of the timeline event
       */
      headerTemplate?: string;
      /**
       * This uses Markdown syntax with Handlebars and event-specific data to render HTML on a timeline when you expand the details.
       */
      detailTemplate?: string;
      /**
       * The tokens of the timeline event
       */
      tokens?: {
        /**
         * The name of the token
         */
        name: string;
        /**
         * The label of the token
         */
        label: string;
        /**
         * The type of the token
         */
        type: "date" | "enumeration" | "number" | "string";
        /**
         * The options of the token if enum
         */
        options?: {
          /**
           * The name of the option
           */
          value: string;
          /**
           * The label of the option
           */
          label: string;
        }[];
        /**
         * The name of the CRM object property. This will populate the CRM object property associated with the event.
         */
        objectPropertyName?: string;
      }[];
      [k: string]: unknown;
    };
  };
}